+++ examples/resources/catalystcenter_floor_image/resource.tf
@@ -1,0 +1,4 @@
+resource "catalystcenter_floor_image" "example" {
+  floor_id    = "34af13e4-da94-4c5f-8da2-aa51a0f0655e"
+  source_path = "testdata/floor_plans/test_floor_plan.png"
+}