+++ internal/provider/provider.go
@@ -347,6 +347,7 @@ func (p *CcProvider) Resources(ctx context.Context) []func() resource.Resource {
 		NewFabricVLANToSSIDResource,
 		NewFabricZoneResource,
 		NewFloorResource,
+		NewFloorImageResource,
 		NewFloorsResource,
 		NewGlobalCredentialNETCONFResource,
 		NewImageResource,